Version 1.1 -Corregir problema en inner class QueryIterator de ExecuteQueryClause, línea 140. Cambiar getInt() por getLong()
- Agregar Types.DECIMAL al mapa DATA_TYPES_MAP de la clase Column.

- agregar parámetro columnExpr al método importModel() para especificar el patrón que se eliminará de las columnas durante la javificación.

-agregar parámetro asPojo al método importModel(), aunque no se implementará.

+++ src/com/fclark/jqlib/clauses/ExecuteQueryClause.java
@@ -137,7 +137,7 @@ public class ExecuteQueryClause extends ExecuteClause  implements Resultable {
                     if (res.getMetaData().getScale(colPos) > 0)
                         result = res.getDouble(colPos);
                     else
-                        result = res.getInt(colPos);
+                        result = res.getLong(colPos);
                 }                
                 break;
             case Types.CHAR:
